refactor(web-awesome): split color scheme into modules
- stores/colorScheme/ → types.ts, constants.ts, store.ts, index.ts
- assets/scss/themes/ → one file per scheme + index.scss
- ColorSchemePicker/ → ColorSchemePicker.tsx, ColorSchemeItem.tsx, index.tsx
- Remove monolithic color-schemes.scss and colorScheme.ts
