+++ packages/web-awesome/src/assets/scss/index.scss
@@ -7,4 +7,4 @@
 @import "../../../../web-components/src/assets/scss/effects.scss";
 @import "../../../../web-components/src/assets/scss/motion.scss";
 @import "common.scss";
-@import "color-schemes.scss";
+@import "themes/index.scss";
