+++ packages/web-awesome/src/assets/scss/themes/_base.scss
@@ -1,0 +1,3 @@
+:root[data-color-scheme] {
+  color-scheme: dark;
+}
