+++ packages/web-awesome/src/assets/scss/themes/index.scss
@@ -1,0 +1,9 @@
+@import "base";
+@import "tokyo-night";
+@import "monokai";
+@import "dracula";
+@import "one-dark";
+@import "catppuccin";
+@import "nord";
+@import "solarized-dark";
+@import "github-dark";
