+++ packages/web-awesome/src/stores/colorScheme/types.ts
@@ -1,0 +1,16 @@
+export type ColorScheme =
+  | "default"
+  | "tokyo-night"
+  | "monokai"
+  | "dracula"
+  | "one-dark"
+  | "catppuccin"
+  | "nord"
+  | "solarized-dark"
+  | "github-dark";
+
+export interface ColorSchemeOption {
+  key: ColorScheme;
+  label: string;
+  preview: string;
+}
